pyproject.toml
[project]
name = "rl-task-contacts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.0.3",
    "pydantic-settings>=2.8.1",
]
